move bootstrap classes to src
