+++ src/Type/Types.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Atk4\Data\Type;
+
+final class Types
+{
+    public const LOCAL_OBJECT = 'atk4_local_object';
+    public const MONEY = 'atk4_money';
+}
