+++ tests/LocalObjectTest.php
@@ -7,8 +7,8 @@ namespace Atk4\Data\Tests;
 use Atk4\Data\Exception;
 use Atk4\Data\Model;
 use Atk4\Data\Schema\TestCase;
-use Atk4\Data\Types\LocalObjectHandle;
-use Atk4\Data\Types\LocalObjectType;
+use Atk4\Data\Type\LocalObjectHandle;
+use Atk4\Data\Type\LocalObjectType;
 use Doctrine\DBAL\Types as DbalTypes;
 
 class LocalObjectTest extends TestCase
